requirements.txt
# Databricks SDK
databricks-sdk>=0.73.0

# PDF Generation
reportlab>=4.0.0

# Email utilities (built-in, but listed for reference)
# smtplib - built-in
# email - built-in

# Configuration management
python-dotenv>=1.0.0

# Optional: Enhanced email capabilities
# sendgrid>=6.10.0  # Uncomment if using SendGrid
# mailgun2>=1.0.0   # Uncomment if using Mailgun

# Optional: Alternative PDF libraries
# weasyprint>=60.0  # HTML to PDF (requires additional system dependencies)
# pdfkit>=1.0.0     # HTML to PDF via wkhtmltopdf (requires wkhtmltopdf installed)

# Development dependencies (optional)
# pytest>=7.4.0
# black>=23.0.0
# mypy>=1.5.0
